TopNav: Move rss feed and kiosk action into profile menu
