+++ public/app/core/components/AppChrome/TopBar/SingleTopBar.tsx
@@ -16,10 +16,10 @@ import { Breadcrumbs } from '../../Breadcrumbs/Breadcrumbs';
 import { buildBreadcrumbs } from '../../Breadcrumbs/utils';
 import { HistoryContainer } from '../History/HistoryContainer';
 import { enrichHelpItem } from '../MegaMenu/utils';
-import { NewsContainer } from '../News/NewsContainer';
 import { QuickAdd } from '../QuickAdd/QuickAdd';
 import { TOP_BAR_LEVEL_HEIGHT } from '../types';
 
+import { ProfileButton } from './ProfileButton';
 import { SignInLink } from './SignInLink';
 import { TopNavBarMenu } from './TopNavBarMenu';
 import { TopSearchBarCommandPaletteTrigger } from './TopSearchBarCommandPaletteTrigger';
@@ -80,23 +80,8 @@ export const SingleTopBar = memo(function SingleTopBar({
             <ToolbarButton iconOnly icon="question-circle" aria-label="Help" />
           </Dropdown>
         )}
-        <ToolbarButton
-          icon="monitor"
-          className={styles.kioskToggle}
-          onClick={onToggleKioskMode}
-          tooltip="Enable kiosk mode"
-        />
         {!contextSrv.user.isSignedIn && <SignInLink />}
-        {profileNode && (
-          <Dropdown overlay={() => <TopNavBarMenu node={profileNode} />} placement="bottom-end">
-            <ToolbarButton
-              className={styles.profileButton}
-              imgSrc={contextSrv.user.gravatarUrl}
-              imgAlt="User avatar"
-              aria-label="Profile"
-            />
-          </Dropdown>
-        )}
+        {profileNode && <ProfileButton profileNode={profileNode} onToggleKioskMode={onToggleKioskMode} />}
       </Stack>
     </div>
   );
@@ -130,15 +115,6 @@ const getStyles = (theme: GrafanaTheme2, menuDockedAndOpen: boolean) => ({
     alignSelf: 'center',
     height: theme.spacing(3),
     width: theme.spacing(3),
-  }),
-  profileButton: css({
-    padding: theme.spacing(0, 0.5),
-    img: {
-      borderRadius: theme.shape.radius.circle,
-      height: '24px',
-      marginRight: 0,
-      width: '24px',
-    },
   }),
   kioskToggle: css({
     [theme.breakpoints.down('lg')]: {
